MUS-33: update tests
