+++ lib/mus-internal/mus-usecase/tests/unittests/test_chat_message_use_case.h
@@ -8,6 +8,7 @@
 
 #include "mus-iusecase/ichat_message_use_case.h"
 #include "mus-irepo/ichat_message_repository.h"
+#include "mus-irepo/ichat_repository.h"
 #include "mus-usecase/chat_message_use_case.h"
 #include "mus-dto/message_request_dto.h"
 #include "mus-dto/message_response_dto.h"
@@ -16,6 +17,8 @@
 #include "mus-usecase/exception/invalid_data_exception.h"
 #include "mus-usecase/exception/null_pointer_exception.h"
 #include "matcher/nullopt_matcher.h"
+#include "mock/mock_chat_repository.h"
+#include "mock/mock_chat_message_repository.h"
 
 using ::testing::AtLeast;
 using testing::Invoke;
@@ -27,6 +30,7 @@ using std::optional;
 using std::shared_ptr;
 using std::string;
 using std::vector;
+using std::make_pair;
 
 using music_share::CreateException;
 using music_share::ExistException;
@@ -35,16 +39,8 @@ using music_share::NullPointerException;
 using music_share::IChatMessageRepository;
 using music_share::ChatMessage;
 using music_share::ChatMessageUseCase;
-
-class MockChatMessageRepository : public IChatMessageRepository {
-public:
-    MOCK_METHOD1(FindByChatId, vector<ChatMessage>(uint32_t));
-    MOCK_METHOD1(Find, optional<ChatMessage>(uint32_t));
-    MOCK_METHOD1(Insert, void(ChatMessage&));
-    MOCK_METHOD1(Update, void(const ChatMessage&));
-    MOCK_METHOD1(Delete, void(const ChatMessage&));
-    MOCK_METHOD2(FindByUserId, vector<ChatMessage>(uint32_t, optional<string>));
-};
+using music_share::IChatRepository;
+using music_share::Chat;
 
 MATCHER_P(ChatMessageEqualement, other,
           "Equality matcher for type ChatMessage") {
@@ -55,7 +51,9 @@ class TestChatMessageUseCase : public ::testing::Test {
 protected:
     void SetUp() {
         chat_message_rep = make_shared<MockChatMessageRepository>();
-        chat_message_usecase = make_shared<ChatMessageUseCase>(*chat_message_rep);
+        chat_rep = make_shared<MockChatRepository>();
+        chat_message_usecase = make_shared<ChatMessageUseCase>(*chat_message_rep,
+                                                               *chat_rep);
         chat_message = make_shared<ChatMessage>(1, "datetime",
                                                 "message", 1, 1);
         message_request = make_shared<MessageRequestDTO>("message");
@@ -65,6 +63,7 @@ protected:
 
     shared_ptr<ChatMessageUseCase> chat_message_usecase;
     shared_ptr<MockChatMessageRepository> chat_message_rep;
+    shared_ptr<MockChatRepository> chat_rep;
     shared_ptr<ChatMessage> chat_message;
     shared_ptr<MessageRequestDTO> message_request;
 };
@@ -110,9 +109,12 @@ TEST_F(TestChatMessageUseCase, GetUserMessagesSuccess) {
                                                  "datetime");
     vector<ChatMessage> messages;
     messages.push_back(*chat_message);
+    Chat chat(make_pair(1, 2), 1);
 
     EXPECT_CALL(*chat_message_rep, FindByChatId(1))
             .WillOnce(Return(messages));
+    EXPECT_CALL(*chat_rep, Find(1))
+            .WillOnce(Return(chat));
 
     vector<MessageResponseDTO> message_response = chat_message_usecase->GetUserMessages(1,
                                                                                         1);
@@ -126,9 +128,12 @@ TEST_F(TestChatMessageUseCase, GetUserMessagesSuccess) {
 
 TEST_F(TestChatMessageUseCase, GetUserMessagesInvalidData) {
     vector<ChatMessage> messages;
+    Chat chat(make_pair(1, 2), 1);
 
     EXPECT_CALL(*chat_message_rep, FindByChatId(1))
             .WillOnce(Return(messages));
+    EXPECT_CALL(*chat_rep, Find(1))
+            .WillOnce(Return(chat));
 
     vector<MessageResponseDTO> messages_response = chat_message_usecase->GetUserMessages(1,
                                                                                          1);
@@ -141,7 +146,10 @@ TEST_F(TestChatMessageUseCase, GetUserMessagesNullPointer) {
     messages.emplace_back(1, "datetime",
                           "message", 1,
                           nullopt);
+    Chat chat(make_pair(1, 2), 1);
 
+    EXPECT_CALL(*chat_rep, Find(1))
+            .WillOnce(Return(chat));
     EXPECT_CALL(*chat_message_rep, FindByChatId(1))
             .WillOnce(Return(messages));
